+++ packages/account-tree-controller/src/rules/entropy.ts
@@ -9,10 +9,10 @@ import { isEvmAccountType } from '@metamask/keyring-api';
 import { KeyringTypes } from '@metamask/keyring-controller';
 import type { InternalAccount } from '@metamask/keyring-internal-api';
 
-import type { AccountGroupObjectOf } from '../group';
-import { BaseRule } from '../rule';
-import type { Rule, RuleResult } from '../rule';
-import type { AccountWalletObjectOf } from '../wallet';
+import type { AccountGroupObjectOf } from '../group.js';
+import { BaseRule } from '../rule.js';
+import type { Rule, RuleResult } from '../rule.js';
+import type { AccountWalletObjectOf } from '../wallet.js';
 
 export class EntropyRule
   extends BaseRule