+++ packages/controller-utils/src/create-service-policy.test.ts
@@ -7,7 +7,7 @@ import {
   DEFAULT_MAX_CONSECUTIVE_FAILURES,
   DEFAULT_MAX_RETRIES,
   ServicePolicy,
-} from './create-service-policy';
+} from './create-service-policy.js';
 
 describe('createServicePolicy', () => {
   beforeEach(() => {