+++ packages/json-rpc-engine/src/createOriginMiddleware.test.ts
@@ -1,8 +1,8 @@
 import { JsonRpcRequest } from '@metamask/utils';
 
-import { makeRequest } from '../tests/utils';
-import { createOriginMiddleware } from './createOriginMiddleware';
-import { JsonRpcEngine } from './JsonRpcEngine';
+import { makeRequest } from '../tests/utils.js';
+import { createOriginMiddleware } from './createOriginMiddleware.js';
+import { JsonRpcEngine } from './JsonRpcEngine.js';
 
 describe('createOriginMiddleware', () => {
   it('adds the origin property to the request', async () => {