+++ packages/money-account-api-data-service/src/index.ts
@@ -3,6 +3,9 @@ export type {
   MoneyAccountApiDataServiceActions,
   MoneyAccountApiDataServiceEvents,
   MoneyAccountApiDataServiceMessenger,
+  MoneyAccountApiDataServiceOptions,
+  MoneyAccountApiDataServiceTraceCallback,
+  MoneyAccountApiDataServiceTraceRequest,
 } from './money-account-api-data-service';
 export type {
   MoneyAccountApiDataServiceFetchPositionsAction,
@@ -12,6 +15,7 @@ export type {
 } from './money-account-api-data-service-method-action-types';
 export type {
   PositionResponse,
+  PositionBalance,
   InterestResponse,
   HistoryResponse,
   RateHistoryResponse,